backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.54.0
pydantic==2.9.2
pydantic[email]==2.9.2
pandas==2.2.3
numpy==2.1.3
httpx>=0.27.0
supabase>=2.8.0
gotrue>=2.10.0
